requirements.txt
pandas
matplotlib
requests
python-binance
python-telegram-bot